Refactor how readings are parsed, add tests
Parsing and filtering the readings in a function called 'readFile' is
not very intuitive and makes it hard to follow what the code is doing.

Refactored `readFile` to only read the file and return its contents.
This data is then parsed and filtered by the callback in
`updateAirQualityIndex` and hopefully made more clear by naming the raw
data "data" and parsed and filtered data "readings".

`readFile`exported for testing purposes.